Changed FlowType from Double to Int
Ich werde bei der Flussberechnung voraussichtlich nur ganze Zahlen brauchen, daher soll die Kapazität auch nur als Integer gespeichert werden

+++ Multicolor-Fair-Clusterin/FairlettBuildupTest.h
@@ -7,7 +7,7 @@ using namespace lemon;
 typedef ListDigraph Graph;
 typedef Graph::Node Node;
 typedef Graph::Arc Arc;
-typedef ListDigraph::ArcMap<double> CapacityMap;
+typedef ListDigraph::ArcMap<int> CapacityMap;
 typedef Preflow<Graph, CapacityMap> Flow;
 
 // Struktur eines Aufgebauten Graphen zum durchreichen durch Methoden
